+++ pkg/ingester/flush.go
@@ -13,6 +13,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prometheus/common/model"
 	"github.com/prometheus/prometheus/pkg/labels"
+	"github.com/prometheus/prometheus/pkg/value"
 
 	"github.com/cortexproject/cortex/pkg/chunk"
 	"github.com/cortexproject/cortex/pkg/util"
@@ -128,6 +129,7 @@ const (
 	reasonMultipleChunksInSeries
 	reasonAged
 	reasonIdle
+	reasonStaleIdle
 )
 
 func (f flushReason) String() string {
@@ -142,6 +144,8 @@ func (f flushReason) String() string {
 		return "Aged"
 	case reasonIdle:
 		return "Idle"
+	case reasonStaleIdle:
+		return "StaleIdle"
 	default:
 		panic("unrecognised flushReason")
 	}
@@ -179,13 +183,13 @@ func (i *Ingester) shouldFlushSeries(series *memorySeries, fp model.Fingerprint,
 		return reasonMultipleChunksInSeries
 	} else if len(series.chunkDescs) > 0 {
 		// Otherwise look in more detail at the first chunk
-		return i.shouldFlushChunk(series.chunkDescs[0], fp)
+		return i.shouldFlushChunk(series.chunkDescs[0], fp, value.IsStaleNaN(float64(series.lastSampleValue)))
 	}
 
 	return noFlush
 }
 
-func (i *Ingester) shouldFlushChunk(c *desc, fp model.Fingerprint) flushReason {
+func (i *Ingester) shouldFlushChunk(c *desc, fp model.Fingerprint, lastValueIsStale bool) flushReason {
 	if c.flushed { // don't flush chunks we've already flushed
 		return noFlush
 	}
@@ -200,9 +204,16 @@ func (i *Ingester) shouldFlushChunk(c *desc, fp model.Fingerprint) flushReason {
 		return reasonAged
 	}
 
-	// Chunk should be flushed if their last update is older then MaxChunkIdle
+	// Chunk should be flushed if their last update is older then MaxChunkIdle.
 	if model.Now().Sub(c.LastUpdate) > i.cfg.MaxChunkIdle {
 		return reasonIdle
+	}
+
+	// A chunk that has a stale marker can be flushed if possible.
+	if i.cfg.MaxStaleChunkIdle > 0 &&
+		lastValueIsStale &&
+		model.Now().Sub(c.LastUpdate) > i.cfg.MaxStaleChunkIdle {
+		return reasonStaleIdle
 	}
 
 	return noFlush
@@ -259,13 +270,13 @@ func (i *Ingester) flushUserSeries(flushQueueIndex int, userID string, fp model.
 
 	// Assume we're going to flush everything, and maybe don't flush the head chunk if it doesn't need it.
 	chunks := series.chunkDescs
-	if immediate || (len(chunks) > 0 && i.shouldFlushChunk(series.head(), fp) != noFlush) {
+	if immediate || (len(chunks) > 0 && i.shouldFlushChunk(series.head(), fp, value.IsStaleNaN(float64(series.lastSampleValue))) != noFlush) {
 		series.closeHead()
 	} else {
 		chunks = chunks[:len(chunks)-1]
 	}
 
-	if reason == reasonIdle && series.headChunkClosed {
+	if (reason == reasonIdle || reason == reasonStaleIdle) && series.headChunkClosed {
 		if minChunkLength := i.limits.MinChunkLength(userID); minChunkLength > 0 {
 			chunkLength := 0
 			for _, c := range chunks {